FILE: renames models.ts to ids.ts
