+++ libs/tree/util/src/lib/shared/types/index.ts
@@ -1,4 +1,4 @@
 export * from './counts';
-export * from './model';
+export * from './ids';
 export * from './node';
 export * from './states';
